+++ web-ui/src/main/resources/catalog/components/metadataactions/RelatedDirective.js
@@ -1195,8 +1195,10 @@
 
   module.directive("gnRecordsTable", [
     "Metadata",
-    "gnRelatedService",
-    function (Metadata, gnRelatedService) {
+    "gnConfigService",
+    "gnConfig",
+    "$rootScope",
+    function (Metadata, gnConfigService, gnConfig, $rootScope) {
       return {
         restrict: "A",
         templateUrl: function (elem, attrs) {
@@ -1217,11 +1219,16 @@
         },
         link: function (scope, element, attrs, controller) {
           var initialized = false;
+          scope.user = $rootScope.user;
           scope.columnsConfig = scope.columns.split(",");
           scope.data = [];
           scope.displayedRecords = [];
           scope.headers = [];
           scope.isArray = angular.isArray;
+
+          gnConfigService.load().then(function (c) {
+            scope.isMdWorkflowEnable = gnConfig["metadata.workflow.enable"];
+          });
 
           if (scope.labels) {
             scope.headers = scope.labels.split(",");
@@ -1230,6 +1237,17 @@
               scope.headers.push(c.startsWith("link/") ? c.split("/")[1] : c);
             });
           }
+
+          scope.isFieldWithRenderer = function (field) {
+            return (
+              field === "resourceType" ||
+              field === "details.stars" ||
+              field.startsWith("link") ||
+              field.startsWith("cl") ||
+              field.startsWith("mdStatus") ||
+              field.startsWith("format")
+            );
+          };
 
           function loadData() {
             scope.data = [];
@@ -1262,7 +1280,7 @@
 
           function sort() {
             scope.displayedRecords.sort(function (a, b) {
-              var sortBy = scope.columnsConfig[0];
+              var sortBy = "resourceTitle";
               return a[sortBy] && a[sortBy].localeCompare(b[sortBy]);
             });
           }